last update struct_cpp

+++ ShteinerAV/Individual task p2/Individual task p2/funcs.cpp
@@ -157,7 +157,7 @@ void fill_vacancy(DbVacancies& Db, const string& link) {
 
 		getline(ss, token, ';');
 		while (token[0] == ' ') token = token.substr(1);
-		Db.arr[current_comp++].salary = stoi(token);
+		Db.arr[current_comp++].salary = stof(token);
 	}
 }
